requirements.txt
ollama>=0.2.1
instagrapi>=2.1.2
linkedin-api>=2.3.1
SQLAlchemy>=2.0.30
schedule>=1.2.1
python-dotenv>=1.0.1
loguru>=0.7.2
fastapi>=0.111.0
uvicorn>=0.30.1
